Refactoring and commenting
Lot's of moving code around to be a bit more readable.  Also made some
functions private that shouldn't be of any use for end users.  Some of
the code got optimized, and maybe a bit easier to read (in my opinion);
but *every-other* was just a useless excerise in optimization as it will
never have a lot of data to work on.