Moved some macro code to real functions
This way there is less stuff happening inside macros, so it's a bit
easier to debug.  With this change, I also moved around a few functions
and renamed form* to transform-validator-forms, which is closer to what
it actually does.  The old defform macro contents is now in the new
form* function.